(REFACTOR) user/home/components 분리 및 폴더 구조화 (#20)
